feat(pt-073): shared compliance runner, wasmtime harness, docs
- Add ai-lib-wasmtime-harness with wasmtime 31 preview1 + correct memory grow math
- Use inline ProtocolManifest-complete YAML (compliance fixtures are looser Value checks)
- Core runs full YAML suite via compliance_from_core; facade compliance reuses runner
- Remove pt073_compliance_subset; refresh README/CHANGELOG

Made-with: Cursor
