+++ Cargo.toml
@@ -5,5 +5,6 @@ members = [
     "crates/ai-lib-contact",
     "crates/ai-lib-rust",
     "crates/ai-lib-wasm",
+    "crates/ai-lib-wasmtime-harness",
 ]
 default-members = ["crates/ai-lib-rust"]
