+++ crates/ai-lib-core/tests/compliance_from_core.rs
@@ -1,0 +1,3 @@
+//! Full YAML compliance suite on **`ai-lib-core`** (same cases as `cargo test --test compliance` on the facade).
+
+mod compliance_runner;
