+++ crates/ai-lib-wasmtime-harness/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ai-lib-wasmtime-harness"
+version = "0.9.3"
+edition = "2021"
+publish = false
+description = "PT-073: wasmtime integration tests for ai-lib-wasm (isolated; heavy deps)"
+
+[dependencies]
+serde_json = "1.0"
+wasmtime = "31"
+wasmtime-wasi = "31"
